[PB-424]:(feature) Fix typeorm entities import
