+++ src/main/database/collections/DriveFolderCollection.ts
@@ -1,0 +1,69 @@
+import { DatabaseCollectionAdapter } from '../adapters/base';
+import { AppDataSource } from '../data-source';
+import { DriveFolder } from '../entities/DriveFolder';
+import { Repository } from 'typeorm';
+export class DriveFoldersCollection
+  implements DatabaseCollectionAdapter<DriveFolder>
+{
+  private repository: Repository<DriveFolder> =
+    AppDataSource.getRepository('drive_folder');
+  async connect(): Promise<{ success: boolean }> {
+    return {
+      success: true,
+    };
+  }
+
+  async get(uuid: DriveFolder['uuid']) {
+    const match = await this.repository.findOneBy({ uuid });
+    return {
+      success: true,
+      result: match,
+    };
+  }
+
+  async getAll() {
+    try {
+      const result = await this.repository.find();
+      return {
+        success: true,
+        result: result,
+      };
+    } catch (error) {
+      return {
+        success: false,
+        result: [],
+      };
+    }
+  }
+
+  async update(uuid: DriveFolder['uuid'], updatePayload: Partial<DriveFolder>) {
+    const match = await this.repository.update(
+      {
+        uuid,
+      },
+      updatePayload
+    );
+
+    return {
+      success: match.affected ? true : false,
+      result: (await this.get(uuid)).result,
+    };
+  }
+
+  async create(creationPayload: DriveFolder) {
+    const createResult = await this.repository.save(creationPayload);
+
+    return {
+      success: createResult ? true : false,
+      result: createResult,
+    };
+  }
+
+  async remove(uuid: DriveFolder['uuid']) {
+    const result = await this.repository.delete({ uuid });
+
+    return {
+      success: result.affected ? true : false,
+    };
+  }
+}
