+++ src/main/database/entities/DriveFile.ts
@@ -1,9 +1,6 @@
-import { DatabaseCollectionAdapter } from '../adapters/base';
-
 import { Column, Entity, PrimaryColumn } from 'typeorm';
-import { AppDataSource } from '../data-source';
 
-@Entity()
+@Entity('drive_file')
 export class DriveFile {
   @Column({ nullable: false })
   fileId!: string;
@@ -51,70 +48,4 @@ export class DriveFile {
 
   @Column({ nullable: false })
   status!: string;
-}
-
-export class DriveFilesCollection
-  implements DatabaseCollectionAdapter<DriveFile>
-{
-  private repository = AppDataSource.getRepository(DriveFile);
-
-  async connect(): Promise<{ success: boolean }> {
-    return {
-      success: true,
-    };
-  }
-
-  async get(uuid: DriveFile['uuid']) {
-    const match = await this.repository.findOneBy({ uuid });
-    return {
-      success: true,
-      result: match,
-    };
-  }
-
-  async getAll() {
-    try {
-      const result = await this.repository.find();
-      return {
-        success: true,
-        result: result,
-      };
-    } catch (error) {
-      return {
-        success: false,
-        result: [],
-      };
-    }
-  }
-
-  async update(uuid: DriveFile['uuid'], updatePayload: Partial<DriveFile>) {
-    const match = await this.repository.update(
-      {
-        uuid,
-      },
-      updatePayload
-    );
-
-    return {
-      success: match.affected ? true : false,
-      result: (await this.get(uuid)).result,
-    };
-  }
-
-  async create(creationPayload: DriveFile) {
-    const createResult = await this.repository.save(creationPayload);
-
-    return {
-      success: createResult ? true : false,
-      result: createResult,
-    };
-  }
-
-  async remove(uuid: DriveFile['uuid']) {
-    const result = await this.repository.delete({ uuid });
-
-    return {
-      success: result.affected ? true : false,
-    };
-  }
 }
