+++ src/main/database/entities/DriveFolder.ts
@@ -1,8 +1,6 @@
-import { DatabaseCollectionAdapter } from '../adapters/base';
 import { Column, Entity, PrimaryColumn } from 'typeorm';
-import { AppDataSource } from '../data-source';
 
-@Entity()
+@Entity('drive_folder')
 export class DriveFolder {
   @PrimaryColumn({
     nullable: false,
@@ -38,69 +36,4 @@ export class DriveFolder {
 
   @Column({ nullable: false })
   status!: string;
-}
-
-export class DriveFoldersCollection
-  implements DatabaseCollectionAdapter<DriveFolder>
-{
-  private repository = AppDataSource.getRepository(DriveFolder);
-  async connect(): Promise<{ success: boolean }> {
-    return {
-      success: true,
-    };
-  }
-
-  async get(uuid: DriveFolder['uuid']) {
-    const match = await this.repository.findOneBy({ uuid });
-    return {
-      success: true,
-      result: match,
-    };
-  }
-
-  async getAll() {
-    try {
-      const result = await this.repository.find();
-      return {
-        success: true,
-        result: result,
-      };
-    } catch (error) {
-      return {
-        success: false,
-        result: [],
-      };
-    }
-  }
-
-  async update(uuid: DriveFolder['uuid'], updatePayload: Partial<DriveFolder>) {
-    const match = await this.repository.update(
-      {
-        uuid,
-      },
-      updatePayload
-    );
-
-    return {
-      success: match.affected ? true : false,
-      result: (await this.get(uuid)).result,
-    };
-  }
-
-  async create(creationPayload: DriveFolder) {
-    const createResult = await this.repository.save(creationPayload);
-
-    return {
-      success: createResult ? true : false,
-      result: createResult,
-    };
-  }
-
-  async remove(uuid: DriveFolder['uuid']) {
-    const result = await this.repository.delete({ uuid });
-
-    return {
-      success: result.affected ? true : false,
-    };
-  }
 }
